refactor: move files related to Mosaic
