+++ giraffe/src/components/Mosaic/index.ts
@@ -1,0 +1,1 @@
+export {MosaicLayer} from './MosaicLayer'
